RaspberryPi_main/Python/requirements.txt
certifi==2019.9.11
chardet==3.0.4
get==2019.4.13
idna==2.8
numpy==1.17.2
post==2019.4.13
public==2019.4.13
PyAudio==0.2.11
pyserial==3.4
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
urllib3==1.25.3
